perf(semdedup): default pairwise precision to fp16

+++ benchmarking/scripts/semdedup_identification_benchmark.py
@@ -47,8 +47,8 @@ def run_semdedup_identification_benchmark(  # noqa: PLR0913
     which_to_keep: str = "hard",
     pairwise_batch_size: int = 1024,
     fit_data_fraction: float | None = None,
-    pairwise_compute_dtype: str = "float32",
-    kmeans_embedding_output_dtype: str = "float32",
+    pairwise_compute_dtype: str = "float16",
+    kmeans_embedding_output_dtype: str = "float16",
     **kwargs,  # noqa: ARG001
 ) -> dict[str, Any]:
     """Run the semantic duplicate identification benchmark and collect comprehensive metrics.
